requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
colorama==0.4.6
coverage==7.4.3
Django==5.0.2
django-debug-toolbar==4.3.0
docopt==0.6.2
Faker==25.2.0
faker_food==0.3.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
parameterized==0.9.0
pillow==10.2.0
pluggy==1.5.0
pyproject==1.3.1
PySocks==1.7.1
pytest==8.0.1
pytest-django==4.8.0
pytest-dotenv==0.5.2
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
watchdog==4.0.0
wsproto==1.2.0
